+++ backends/carp_webservices/test/carp_deployment_service_test.dart
@@ -1,5 +1,4 @@
 import 'dart:convert';
-import 'dart:math';
 
 import 'package:carp_mobile_sensing/carp_mobile_sensing.dart';
 import 'package:carp_webservices/carp_auth/carp_auth.dart';
@@ -22,13 +21,10 @@ void main() {
   /// Setup CARP and authenticate.
   /// Runs once before all tests.
   setUpAll(() async {
-    study = new Study(
-        id: testStudyId,
-        userId: userId,
-        deploymentId: testDeploymentId,
-        name: "Test study");
+    study = new Study(id: testStudyId, userId: userId, name: "Test study");
     app = new CarpApp(
       study: study,
+      studyDeploymentId: testDeploymentId,
       name: "Test",
       uri: Uri.parse(uri),
       oauth: OAuthEndPoint(clientID: clientID, clientSecret: clientSecret),
@@ -71,11 +67,12 @@ void main() {
       print(_encode(status.toJson()));
       print(status);
       print(status.masterDeviceStatus.device);
-      expect(status.studyDeploymentId, study.deploymentId);
+      expect(status.studyDeploymentId, testDeploymentId);
     }, skip: false);
 
     test('- register device', () async {
-      DeploymentReference reference = CarpService().deployment();
+      DeploymentReference reference =
+          CarpService().deployment(testDeploymentId);
       StudyDeploymentStatus status = await reference.getStatus();
       print(status);
       expect(status.masterDeviceStatus.device, isNotNull);
@@ -83,32 +80,39 @@ void main() {
       status = await reference.registerDevice(
           deviceRoleName: status.masterDeviceStatus.device.roleName);
       print(status);
-      expect(status.studyDeploymentId, study.deploymentId);
+      expect(status.studyDeploymentId, testDeploymentId);
     }, skip: false);
 
     test('- get master device deployment', () async {
-      DeploymentReference reference = CarpService().deployment();
+      DeploymentReference reference =
+          CarpService().deployment(testDeploymentId);
       StudyDeploymentStatus status = await reference.getStatus();
       print(status);
       expect(status.masterDeviceStatus.device, isNotNull);
       print(status.masterDeviceStatus.device);
       MasterDeviceDeployment deployment = await reference.get();
       print(deployment);
+      deployment.tasks.forEach((task) {
+        print(task);
+        task?.measures?.forEach(print);
+      });
       expect(deployment.configuration.deviceId, isNotNull);
     }, skip: false);
 
     test('- deployment success', () async {
-      DeploymentReference reference = CarpService().deployment();
+      DeploymentReference reference =
+          CarpService().deployment(testDeploymentId);
       StudyDeploymentStatus status_1 = await reference.getStatus();
       MasterDeviceDeployment deployment = await reference.get();
       StudyDeploymentStatus status_2 = await reference.success();
       print(status_2);
       expect(status_1.studyDeploymentId, status_2.studyDeploymentId);
-      expect(status_2.studyDeploymentId, study.deploymentId);
+      expect(status_2.studyDeploymentId, testDeploymentId);
     }, skip: false);
 
     test('- unregister device', () async {
-      DeploymentReference reference = CarpService().deployment();
+      DeploymentReference reference =
+          CarpService().deployment(testDeploymentId);
       StudyDeploymentStatus status = await reference.getStatus();
       print(status);
       expect(status.masterDeviceStatus.device, isNotNull);
@@ -116,7 +120,7 @@ void main() {
       status = await reference.unRegisterDevice(
           deviceRoleName: status.masterDeviceStatus.device.roleName);
       print(status);
-      expect(status.studyDeploymentId, study.deploymentId);
+      expect(status.studyDeploymentId, testDeploymentId);
     }, skip: false);
   }, skip: true);
 }