+++ carp_mobile_sensing/lib/data_managers/local/file_study_manager.dart
@@ -75,6 +75,4 @@ class FileStudyManager implements StudyManager {
   ///   `carp/study/study-<study_id>.json`
   ///
   String filename(String studyId) => '$_path/study-$studyId.json';
-
-  Future dispose() {}
 }