+++ carp_mobile_sensing/test/core_json_test.dart
@@ -16,10 +16,10 @@ void main() {
     //SamplingPackageRegistry.register(ContextSamplingPackage());
 
     study = Study(
-        id: '1234',
-        userId: 'bardram',
-        name: 'bardram study',
-        deploymentId: '#1');
+      id: '1234',
+      userId: 'bardram',
+      name: 'bardram study',
+    );
     //study.dataEndPoint = DataEndPoint(DataEndPointType.PRINT);
     study.dataEndPoint = FileDataEndPoint()
       ..bufferSize = 50 * 1000
@@ -30,8 +30,7 @@ void main() {
     study.addTriggerTask(
         ImmediateTrigger(), // a simple trigger that starts immediately
         AutomaticTask(name: 'Sampling Task')
-          ..measures = SamplingSchema
-              .common(namespace: NameSpace.CARP)
+          ..measures = SamplingSchema.common(namespace: NameSpace.CARP)
               .measures
               .values
               .toList() // a task with all measures
@@ -42,7 +41,6 @@ void main() {
     print(_encode(study));
 
     expect(study.id, '1234');
-    expect(study.deploymentId, '#1');
   });
 
   test('JSON -> Study, assert study id', () async {
@@ -51,7 +49,6 @@ void main() {
     Study study_2 =
         Study.fromJson(json.decode(studyJson) as Map<String, dynamic>);
     expect(study_2.id, study.id);
-    expect(study.deploymentId, '#1');
 
     print(_encode(study_2));
   });
@@ -136,8 +133,7 @@ void main() {
     study_3.addTriggerTask(
         t1,
         AutomaticTask(name: 'Sensing Task #1')
-          ..measures = SamplingSchema
-              .common()
+          ..measures = SamplingSchema.common()
               .getMeasureList(types: [DeviceSamplingPackage.MEMORY]));
 
     // collect every other day at 13:30.
@@ -190,8 +186,7 @@ void main() {
                 MeasureType(NameSpace.CARP, DeviceSamplingPackage.BATTERY),
             resumeCondition: BatteryDatum()..batteryLevel = 10),
         AutomaticTask(name: 'Sensing Task #1')
-          ..measures = SamplingSchema
-              .common()
+          ..measures = SamplingSchema.common()
               .getMeasureList(types: [SensorSamplingPackage.LIGHT]));
 
     study_3.addTriggerTask(
@@ -201,8 +196,7 @@ void main() {
             resumeCondition: (datum) =>
                 (datum as BatteryDatum).batteryLevel == 10),
         AutomaticTask(name: 'Sensing Task #1')
-          ..measures = SamplingSchema
-              .common()
+          ..measures = SamplingSchema.common()
               .getMeasureList(types: [SensorSamplingPackage.LIGHT]));
 
     final studyJson = _encode(study_3);